Duplicate code to util file
