+++ src/integration/kotlin/Kafka2hbUcfsIntegrationSpec.kt
@@ -61,7 +61,7 @@ class Kafka2hbUcfsIntegrationSpec : StringSpec() {
             val summariesManifests1 = s3Client.listObjectsV2("manifests", "manifest_prefix").objectSummaries
             summariesManifests1.size shouldBe 1
 
-            verifyHbaseRegions()
+            verifyHbaseRegions(expectedTablesToRegions, regionReplication, regionServers)
             verifyMetadataStore(1, topic, true)
         }
 
@@ -133,7 +133,6 @@ class Kafka2hbUcfsIntegrationSpec : StringSpec() {
                 waitFor { hbase.getCellBeforeTimestamp(qualifiedTableName, hbaseKey, referenceTimestamp) }
             String(storedPreviousValue!!) shouldBe String(body1)
 
-//            verifyHbaseRegions()
             verifyMetadataStore(1, topic, true)
         }
 
@@ -230,47 +229,6 @@ class Kafka2hbUcfsIntegrationSpec : StringSpec() {
             summaries1.size shouldBe 0
             verifyMetadataStore(1, topic, true)
         }
-    }
-
-    private suspend fun verifyHbaseRegions() {
-
-        var waitSoFarSecs = 0
-        val longInterval = 5
-
-        logger.info("Waiting for $expectedTablesToRegions hbase tables to appear with given regions")
-
-        val foundTablesToRegions = mutableMapOf<String, Int>()
-
-        HbaseClient.connect().use { hbase ->
-            withTimeout(10.minutes) {
-                do {
-                    val foundTablesSorted = testTables()
-                    logger.info("Waiting for $expectedTablesToRegions hbase tables to appear",
-                        "found_tables_so_far", "${foundTablesSorted.size}",
-                        "total_seconds_elapsed", "$waitSoFarSecs")
-                    delay(longInterval.seconds)
-                    waitSoFarSecs += longInterval
-                }
-                while (expectedTablesToRegions != foundTablesSorted.size)
-
-                testTables().forEach { tableName ->
-                    launch(Dispatchers.IO) {
-                        val regionsWithReplication = hbase.getTableRegions(TableName.valueOf(tableName)).size
-                        logger.info(
-                            "Found table",
-                            "table_name", tableName,
-                            "regions_with_replication", "$regionsWithReplication",
-                        )
-                        foundTablesToRegions[tableName] = regionsWithReplication
-                    }
-                }
-            }
-        }
-
-        logger.info(foundTablesToRegions.toString())
-
-        foundTablesToRegions.values.contains(regionReplication * regionServers) shouldBe true
-        foundTablesToRegions.values.toSet().size shouldBe 1
     }
 }
